block-explorer-ui/libs/state/transactions-state/src/lib/+state/transactions.facade.ts
import { Injectable } from '@angular/core';

import { select, Store } from '@ngrx/store';

import { TransactionsPartialState } from './transactions.reducer';
import { transactionsQuery } from './transactions.selectors';
import { LoadTransactions } from './transactions.actions';

@Injectable()
export class TransactionsFacade {
  loaded$ = this.store.pipe(select(transactionsQuery.getLoaded));
  allTransactions$ = this.store.pipe(
    select(transactionsQuery.getAllTransactions)
  );
  selectedTransactions$ = this.store.pipe(
    select(transactionsQuery.getSelectedTransactions)
  );

  constructor(private store: Store<TransactionsPartialState>) {}

  loadAll() {
    this.store.dispatch(new LoadTransactions());
  }
}
